maybe reference is a better name?
